Split up Analyzer.
